Consolidate string falsiness, JS trim, and expires_in modeling onto shared helpers
- truthyString adoption: JSON-object sites whose upstream is a JS truthiness
  read now declare it (completions/mistral finish_reason + responseId, google
  responseId/thoughtSignature, responses-shared arguments, codex event type /
  error code||type / plan_type / message, app tool-call titles + brave
  description); sites with typeof/length or null/undefined upstreams keep
  their existing spellings.
- One shared MAX_SAFE_INTEGER in ai/utils/JsNumbers.kt; SettingsManager and
  OpenRouterOAuthAuth use it.
- Device-flow expires_in modeled as a JS number end to end: the poller
  deadline stays a Double (fractional survives; Infinity never wraps like the
  old Long*1000 overflow), sleeps convert through Node setTimeout semantics
  (trunc, sub-1ms clamp); Kimi drops the Long roundtrip, Copilot/Xai/Codex
  pass Doubles; xai refresh epoch computed in doubles before the Long store.
- JS trim set applied at every site whose upstream trims: EditDiff per-line
  trimEnd (U+FEFF now stripped, U+001C-U+001F kept, like JS), provider error
  bodies, partial-json blank/parse, Mistral SSE/body/tool-result/text,
  codex SSE data + base-url gate, anthropic beta-features/content/thinking,
  azure deployment map/base-URLs (also fixes the `||` fallthrough for empty
  overrides), constrained-sampling grammar guards, OAuth paste inputs,
  model-reference parsing, image data-URL mime, ModelsError detail, and the
  Date.parse stand-in's whitespace trim (V8 DateParser set).
- sanitizeSurrogates: verified pi's replay path sanitizes neither the raw
  reasoning field nor reasoning_details; test comments now cite the upstream
  and a reasoning_details-verbatim pin was added.
- Cross-referenced the intentional formQuery (lenient) vs PathUtils
  percentDecode (strict) asymmetry in both KDocs.

+++ packages/ai/src/main/kotlin/works/resolve/pathfinder/ai/auth/oauth/KimiCodingOAuthAuth.kt
@@ -63,7 +63,7 @@ class KimiCodingOAuthAuth(
         val verificationUri: String,
         val verificationUriComplete: String,
         val intervalSeconds: Double,
-        val expiresInSeconds: Long
+        val expiresInSeconds: Double
     )
 
     internal data class TokenResponse(val access: String, val refresh: String, val expires: Long)
@@ -118,9 +118,9 @@ class KimiCodingOAuthAuth(
             expiresInSeconds = if (expiresIn != null && expiresIn.isFinite() &&
                 expiresIn > 0
             ) {
-                expiresIn.toLong()
+                expiresIn
             } else {
-                DEVICE_CODE_TIMEOUT_SECONDS
+                DEVICE_CODE_TIMEOUT_SECONDS.toDouble()
             }
         )
     }
